+++ src/services/appliers/TranslationApplyCommand.php
@@ -32,17 +32,28 @@ class TranslationApplyCommand
     private $targetLanguage;
 
     /**
+     * @var TranslationRecord
+     */
+    private $translationRecord;
+
+    /**
      * @param ElementInterface $element
      * @param Job $job
      * @param array $content
      * @param string $targetLanguage
      */
-    public function __construct(ElementInterface $element, Job $job, array $content, string $targetLanguage)
-    {
+    public function __construct(
+        ElementInterface $element,
+        Job $job,
+        array $content,
+        string $targetLanguage,
+        TranslationRecord $translationRecord
+    ) {
         $this->element = $element;
         $this->job = $job;
         $this->content = $content;
         $this->targetLanguage = $targetLanguage;
+        $this->translationRecord = $translationRecord;
     }
 
     /**
@@ -85,5 +96,13 @@ class TranslationApplyCommand
     public function getSourceSiteId(): int
     {
         return $this->job->sourceSiteId;
+    }
+
+    /**
+     * @return TranslationRecord
+     */
+    public function getTranslationRecord(): TranslationRecord
+    {
+        return $this->translationRecord;
     }
 }